+++ tour-service/src/main/java/com/example/tourservice/TravelAgentService.java
@@ -1,4 +1,6 @@
-package com.example.tourservice.utilities;
+package com.example.tourservice;
+
+import com.example.tourservice.utilities.TourRepository;
 
 public class TravelAgentService {
     private  TourRepository tourRepository;
